+++ fractal_server/tasks/v2/local/collect_pixi.py
@@ -1,4 +1,5 @@
 import json
+import os
 import shutil
 import time
 from pathlib import Path
@@ -90,14 +91,17 @@ def collect_local_pixi(
 
                 logger.info(f"{use_pixi_lockfile=}")
                 frozen_option = "--frozen" if use_pixi_lockfile else ""
+                pixi_home = resource.tasks_pixi_config["versions"][
+                    task_group.pixi_version
+                ]
+                pixi_cache_dir = profile.pixi_cache_dir or os.path.join(
+                    pixi_home, "cache"
+                )
+                logger.info(f"Setting PIXI_CACHE_DIR to {pixi_cache_dir}")
                 common_args = dict(
                     replacements={
-                        (
-                            "__PIXI_HOME__",
-                            resource.tasks_pixi_config["versions"][
-                                task_group.pixi_version
-                            ],
-                        ),
+                        ("__PIXI_HOME__", pixi_home),
+                        ("__PIXI_CACHE_DIR__", pixi_cache_dir),
                         ("__PACKAGE_DIR__", task_group.path),
                         ("__TAR_GZ_PATH__", archive_path),
                         (