+++ fractal_server/tasks/v2/templates/pixi_3_post_install.sh
@@ -9,6 +9,7 @@ write_log(){
 
 # Replacements
 PIXI_HOME="__PIXI_HOME__"
+PIXI_CACHE_DIR="__PIXI_CACHE_DIR__"
 PACKAGE_DIR="__PACKAGE_DIR__"
 SOURCE_DIR_NAME="__SOURCE_DIR_NAME__"
 IMPORT_PACKAGE_NAME="__IMPORT_PACKAGE_NAME__"
@@ -26,7 +27,7 @@ PROJECT_PYTHON_WRAPPER="${SOURCE_DIR}/project_python.sh"
 
 # Pixi env variable
 export PIXI_HOME="${PIXI_HOME}"
-export PIXI_CACHE_DIR="${PIXI_HOME}/cache"
+export PIXI_CACHE_DIR="${PIXI_CACHE_DIR}"
 export RATTLER_AUTH_FILE="${PIXI_HOME}/credentials.json"